fix(myTravel): move the travel page
